refactor(cockpit): extract reusable MarketTeardown so a market is data, not a page
Markets 3-8 should not each need a bespoke Vue page. Extracts the proven market-2
surface into components/MarketTeardown.vue driven by a shared MarketTeardownData
type, and reduces the market-2 page to a 10-line wrapper.

Each additional market is now a data module plus a thin page, which is what makes
covering the remaining six markets tractable rather than a rewrite each time.

- features/competitive-intelligence/marketTeardown.ts: shared types
  (TeardownSpecimen / TeardownMachineStep / TeardownGap / MarketTeardownData).
  moatArchetype is free-text per market, because each market names its own moat
  kinds — enterprise 'permissions graph' has no analogue in browsers.
- components/MarketTeardown.vue: the surface (header, boundary, thesis, our read,
  machine, durability-ranked specimen table, unserved gaps).
- marketProfessionalIntelligence.ts now types as MarketTeardownData.

No behaviour or content change; market 2 renders identically.

verify: typecheck + 543 tests (103 files) + build all green.
